refactor(cli,core): keep @jridgewell/trace-mapping out of @pikku/core
The coverage span mapping moves to the CLI. @pikku/core keeps only the
report types plus an optional CoverageService.takeReport, which the
CLI-booted coverage service implements — the scaffolded
pikkuScenarioTakeLiveCoverage function reaches the mapping through the
injected service instead of importing it from core.
