+++ .changeset/scaffold-scenarios.md
@@ -15,6 +15,9 @@ installed.
 A new `scaffold.scenarios` feature (`pikku enable scenarios`, or
 `"scaffold": { "scenarios": "auth" }` in pikku.config.json) generates the
 four functions into the project scaffold as `pikkuScenario*` exposed RPCs.
-The scenario runner and `expectService` now invoke those names, the coverage
-mapping helpers moved from the addon into `@pikku/core/services`, and the
-addon copies were removed.
+The scenario runner and `expectService` now invoke those names and the
+addon copies were removed. The source-map-aware coverage mapping (and its
+`@jridgewell/trace-mapping` dependency) moved from the addon into the CLI:
+`@pikku/core` gains only the report types plus an optional
+`CoverageService.takeReport`, which the CLI-booted coverage service
+implements and the scaffolded function calls.
