+++ packages/core/src/services/index.ts
@@ -162,17 +162,12 @@ export {
   type ScriptCoverage,
   type FunctionCoverage,
   type CoverageRange,
-} from './v8-coverage-service.js'
-export { IstanbulCoverageService } from './istanbul-coverage-service.js'
-export {
-  mapPreciseCoverage,
-  mapLineHitsToReport,
   type CoverageStatus,
   type FunctionCoverageEntry,
   type FunctionCoverageReport,
   type CoverageFunctionMeta,
-  type ScriptCoverageInput,
-} from './precise-coverage-map.js'
+} from './v8-coverage-service.js'
+export { IstanbulCoverageService } from './istanbul-coverage-service.js'
 export {
   StubTracker,
   createStubProxy,
